requirements.txt
# Seoul Safety Data Pipeline - Required Dependencies

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# PostgreSQL database support
psycopg2-binary>=2.9.0

# HTTP requests and API calls
requests>=2.28.0

# Environment variables management
python-dotenv>=0.19.0

# Geospatial data processing (optional - for advanced features)
geopandas>=0.12.0
shapely>=1.8.0
pyproj>=3.4.0
rtree>=1.0.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and logging
tqdm>=4.64.0

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0